pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lfw-verification"
version = "0.1.0"
description = "Reproducible LFW ingestion + verification pair generation + vectorized similarity scoring."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.22",
  "pandas>=1.5",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
  "ruff>=0.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
